# internal/app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: fs-access-api
  version: "1.0"
  description: >
    Administrative API for managing filesystem access entities such as users and groups.<br>
    All non-public endpoints require authentication using either the **HMAC** or **Bearer** scheme (depending on the configuration).<br>
    The **Bearer** scheme requires the headers `X-Api-Key` and `Authorization`.<br>
    The **HMAC** scheme requires the headers `X-Api-Key`, `Authorization`, `X-Timestamp`, and `X-Content-Sha256`.
servers:
  - url: /
components:
  parameters:
...
    UsernameParam:
      name: username
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Username' }
      description: Resource identifier (username). Slash not allowed.
    DirnameParam:
      name: dirname
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Dirname' }
      description: Resource identifier (dirname). Slash not allowed.
  headers:
    LocationHeader:
      description: URL of the created resource
      schema: { type: string, format: uri }
  securitySchemes:
    XApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: API key identifier used to select the matching secret.
    Authorization:
      type: apiKey
      in: header
      name: authorization
      description: >
        For HMAC authentication: use `Authorization: HMAC <hex-signature>`  
        For Bearer authentication: use `Authorization: Bearer <api-secret>`
    XTimestamp:
      type: apiKey
      in: header
      name: x-timestamp
      description: >
        For HMAC authentication: must be `timestamp=<unix-timestamp>`  
        (used to prevent replay attacks).  
        Omit this header when using Bearer authentication.
    XContentSha256:
      type: apiKey
      in: header
      name: x-content-sha256
      description: >
        For HMAC authentication: must be `sha256=<hex-digest>`  
        (the SHA-256 hash of the request body).  
        Omit this header when using Bearer authentication.
  responses:
    Unauthorized:
      description: Missing/invalid authentication
    Forbidden:
      description: Insufficient permissions
...
    Created:
      description: Created
      headers:
        Location:
          $ref: '#/components/headers/LocationHeader'
    Updated:
      description: Updated
      headers:
        Location:
          $ref: '#/components/headers/LocationHeader'
    Deleted:
      description: Deleted
    NoContent:
      description: No content
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict — resource exists with different data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        code: { type: string, example: USER_CONFLICT }
        message: { type: string }
      required: [ code, message ]
    RelativePath:
      type: string
      nullable: false
      minLength: 1
      maxLength: 1024
      pattern: '^[A-Za-z0-9._-][A-Za-z0-9._/-]*$'
      description: >
        Relative path string that must not start with a slash (`/`) and must not
        contain spaces. Allowed characters: letters, digits, dot (`.`), underscore (`_`),
        hyphen (`-`), and slash (`/`).  
    Description:
      type: string
      nullable: true
      maximum: 255
    UID:
      type: integer
      nullable: false
      format: uint32
      minimum: 2000
      maximum: 4294967295
    GID:
      type: integer
      nullable: false
      format: uint32
      minimum: 2000
      maximum: 4294967295
    Groupname:
      type: string
      nullable: false
      minimum: 2
      maximum: 128
      pattern: '^[A-Za-z0-9._-]+$'
      description: Group name. Slash (/) is not allowed.
    Username:
      type: string
      nullable: false
      minimum: 2
      maximum: 128
      pattern: '^[A-Za-z0-9._-]+$'
      description: Username. Slash (/) is not allowed.
    Dirname:
      type: string
      nullable: false
      minimum: 2
      maximum: 128
      pattern: '^[A-Za-z0-9._-]+$'
      description: Directory name. Slash (/) is not allowed.
...
    UserInfo:
      type: object
      additionalProperties: false
      required: [ username, uid, description, groupname, gid, home, disabled ]
      properties:
        username: { $ref: '#/components/schemas/Username' }
        uid: { $ref: '#/components/schemas/UID' }
        description: { $ref: '#/components/schemas/Description' }
        groupname: { $ref: '#/components/schemas/Groupname' }
        gid: { $ref: '#/components/schemas/GID' }
        home: { $ref: '#/components/schemas/RelativePath' }
        expiration: { type: string, format: date-time, nullable: true }
        disabled: { type: boolean }
    EnsureUserRequestBody:
      type: object
      additionalProperties: false
      required: [ groupname, password, password_is_hash ]
      properties:
        description: { $ref: '#/components/schemas/Description' }
        groupname: { $ref: '#/components/schemas/Groupname' }
        home: { $ref: '#/components/schemas/RelativePath' }
        expiration: { type: string, format: date-time, nullable: true }
        disabled: { type: boolean, default: false }
        password:
          type: string
          writeOnly: true
          minLength: 8
          description: >
            Plaintext or final hash depending on `password_is_hash`.
        password_is_hash:
          type: boolean
          writeOnly: true
          description: >
            When true, `password` is treated as a final hash; otherwise it will be hashed server-side.
    SetDescriptionRequestBody:
      type: object
      additionalProperties: false
      required: [ description ]
      properties:
        description: { $ref: '#/components/schemas/Description' }
    SetUserPasswordRequestBody:
      type: object
      additionalProperties: false
      required: [ password, password_is_hash ]
      properties:
        password:
          type: string
          writeOnly: true
          minLength: 8
          description: >
            User password input. If `password_is_hash=false`, the server will hash
            it using the default algorithm. If `password_is_hash=true`, the value
            must already be a final hash.
        password_is_hash:
          type: boolean
          writeOnly: true
          description: >
            When true, `password` is treated as a final hash value. When false,
            the server will hash the given plaintext password before storing it.
    SetUserExpirationRequestBody:
      type: object
      additionalProperties: false
      required: [ expiration ]
      properties:
        expiration: { type: string, format: date-time, nullable: true }
    SetUserDisabledRequestBody:
      type: object
      additionalProperties: false
      required: [ disabled ]
      properties:
        disabled: { type: boolean }
security:
  - XApiKey: [ ]
    Authorization: [ ]
    XTimestamp: [ ]
    XContentSha256: [ ]
paths:
...
  /api/users:
    get:
      operationId: ListUsers
      summary: List users (without passwords)
      tags: [ Users ]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInfo'
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /api/users/{username}:
    parameters:
      - $ref: '#/components/parameters/UsernameParam'
    get:
      operationId: GetUser
      summary: Get user details (without password)
      tags: [ Users ]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserInfo' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }

    put:
      operationId: EnsureUser
      summary: Create-or-ensure user (idempotent)
      description: |
        Ensures the user identified by `{username}` exists with the requested state.
      tags: [ Users ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnsureUserRequestBody' }
      responses:
        '200': { $ref: '#/components/responses/Updated' }
        '201': { $ref: '#/components/responses/Created' }
        '409': { $ref: '#/components/responses/Conflict' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }

    delete:
      operationId: DeleteUser
      summary: Delete user
      tags: [ Users ]
      responses:
        "204": { $ref: '#/components/responses/Deleted' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /api/users/{username}/description:
    parameters:
      - $ref: '#/components/parameters/UsernameParam'
    put:
      operationId: SetUserDescription
      summary: Set or change user description
      tags: [ Users ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetDescriptionRequestBody' }
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /api/users/{username}/password:
    parameters:
      - $ref: '#/components/parameters/UsernameParam'
    put:
      operationId: SetUserPassword
      summary: Set or change user password
      description: Sets new password without requiring the previous one.
      tags: [ Users ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetUserPasswordRequestBody' }
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /api/users/{username}/expiration:
    parameters:
      - $ref: '#/components/parameters/UsernameParam'
    put:
      operationId: SetUserExpiration
      summary: Set or change user expiration
      tags: [ Users ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetUserExpirationRequestBody' }
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /api/users/{username}/disabled:
    parameters:
      - $ref: '#/components/parameters/UsernameParam'
    put:
      operationId: SetUserDisabled
      summary: Set or change user disabled status
      tags: [ Users ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetUserDisabledRequestBody' }
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /api/users/{username}/directories:
    parameters:
      - $ref: '#/components/parameters/UsernameParam'
    get:
      operationId: ListUserDirs
      summary: List user top-level directories
      tags: [ Directories ]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dirname'
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /api/users/{username}/directories/{dirname}:
    parameters:
      - $ref: '#/components/parameters/UsernameParam'
      - $ref: '#/components/parameters/DirnameParam'

    put:
      operationId: EnsureUserDir
      summary: Create-or-ensure user directory (idempotent)
      description: |
        Ensures the user's top-level directory identified by `{dirname}` exists with the requested state.
      tags: [ Directories ]
      responses:
        '200': { $ref: '#/components/responses/Updated' }
        '201': { $ref: '#/components/responses/Created' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }

    delete:
      operationId: DeleteUserDir
      summary: Delete user top-level directory if doesn't contain any files
      tags: [ Directories ]
      responses:
        "204": { $ref: '#/components/responses/Deleted' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...